Add error handling in Lookups.
Problem resolved: without error-handling in lookups we have to assume that the
server replies with a proper response for every query made. This passes server
errors on as client errors. While normally this would be a library concern it'd
better to give the client a first-order way of handling these errors as opposed
to having incomplete pattern matches like we had before.

The solution applied here builds a general error handling system into response
parsing. This is also appropriate for later when we want to actually report
server errors to the client.