Simplify QueryMode and eliminate need for .get
Problem: QueryMode introduces an unneeded layer of indirection, reducing
efficiency and adding boilerplate to the interface.

We can squash this indirection out by having HasQuery define the lookup values
directly. This also starts to aid in some of the naming difficulties we've faced
so far. Finally, it eliminates the need to use .get when building queries and
makes the meaning of Lookup more clear.