Nest LookupError to mark Unexpected errors
Problem: It's not obvious to the user what a ResponseMissing error means.

ResponseMissing (or ResponseMissingCRITICAL as it was named) occurs when there
is a bug in the library and a bad Lookup value has been created. Eventually,
once library confidence is high, this will become an exception and will be
impossible for the user to handle all-together. In the mean time, we construct a
new class of LookupErrors called Unexpected(..) to make this case clear.

Additionally, we rename the error to ServerShouldHaveResponded. Stupid-long but
also becoming increasingly clear.