Support indexing of queries
Problem: Interesting queries are parameterized or indexed by a particular
request from the client. Without support for indexes, these queries are
necessarily quite Useless!

This patch adds IndexedAtom and IndexedObj which are like normal Atoms and Objs,
but they support providing an "index value" at query-building time which
parameterizes the request.

Squashed commit of the following:

commit aedac6875508f452b0a6d40b7d5c5f229c2ad7f6
Author: Joseph Abrahamson <me@jspha.com>
Date:   Mon Mar 6 18:42:56 2017 -0500

    Add IndexedObj

commit 308dd7ec7e30b34e7ab3542db10b1533f53a27a1
Author: Joseph Abrahamson <me@jspha.com>
Date:   Mon Mar 6 18:07:57 2017 -0500

    Differentiate between IndexedAtom and Atom

commit 616298cac7fa27122a0d09e45dd2caf8abc52ddd
Author: Joseph Abrahamson <me@jspha.com>
Date:   Mon Mar 6 17:49:40 2017 -0500

    Add arguments for Atomic types