Remove Multiplicity#Wrap
Problem: Multiplicity#Wrap induces a layer of excess indirection. This means
that Obj[A] and ObjM[Singular, A] aren't technically identical. Additionally, it
leads to a little extra line noise.

Previously I thought Wrap was necessary so that the implicit resolution had
something concrete to "work backwards" from. Turns out this isn't the
case---instead I just have to generate Interprets instances for each choice of
multiplicity so matching occurs directly.